+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (Unreleased)
+
 - Add env variable lookup for k8s client settings (https://github.com/pulumi/pulumi-kubernetes/pull/1777)
+- Fix diff logic for CRD kinds with the same name as a built-in (https://github.com/pulumi/pulumi-kubernetes/pull/1779)
 
 ## 3.8.2 (October 18, 2021)